+++ ffi/src/main/kotlin/gay/pizza/pork/ffi/FfiType.kt
@@ -1,5 +1,5 @@
 package gay.pizza.pork.ffi
 
 interface FfiType {
-  val size: Int
+  val size: Long
 }
